feat(cmd/wallet): Refactor wallet cmd into separate package
