+++ cmd/root.go
@@ -13,6 +13,7 @@ import (
 	"github.com/oasisprotocol/cli/cmd/account"
 	"github.com/oasisprotocol/cli/cmd/network"
 	"github.com/oasisprotocol/cli/cmd/paratime"
+	"github.com/oasisprotocol/cli/cmd/wallet"
 	"github.com/oasisprotocol/cli/config"
 	"github.com/oasisprotocol/cli/version"
 	_ "github.com/oasisprotocol/cli/wallet/file"   // Register file wallet backend.
@@ -98,7 +99,7 @@ func init() {
 
 	rootCmd.AddCommand(network.Cmd)
 	rootCmd.AddCommand(paratime.Cmd)
-	rootCmd.AddCommand(walletCmd)
+	rootCmd.AddCommand(wallet.Cmd)
 	rootCmd.AddCommand(account.Cmd)
 	rootCmd.AddCommand(addressBookCmd)
 	rootCmd.AddCommand(contractCmd)
